src/main/java/hu/progmatic/kozos/kassza/TermekSzerviz.java
package hu.progmatic.kozos.kassza;

import org.springframework.beans.factory.InitializingBean;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.Collection;
import java.util.List;

@Service
@Transactional
public class TermekSzerviz  {


    @Autowired
    private TermekRepository repository;


    public Termek save(Termek termek) {
        return repository.saveAndFlush(termek);
    }

    public Termek getById(Integer id) {
        return repository.getById(id);
    }

    public void saveAll(List<Termek> initItems) {
        repository.saveAllAndFlush(initItems);
    }

    public List<Termek> findAllbyNev(String nev) {
        return repository.findAllByMegnevezesContains(nev);
    }

    public List<Termek> findAll() {
        return repository.findAll();
    }

    public Termek getByVonalkod(String vonalkod){
        return repository.findByVonalkod(vonalkod);
    }


}
